refactor: update i18n types and config
