+++ astro.config.mjs
@@ -25,6 +25,7 @@ import remarkCollapse from 'remark-collapse';
 import numberedFootnoteLabels from 'remark-numbered-footnote-labels';
 import remarkToc from 'remark-toc';
 
+import { defaultLocale, explicitLocales, locales } from './src/i18n/config';
 import HashRenamer from './src/lib/hash-renamer';
 import getRedirects from './src/lib/redirects';
 
@@ -45,16 +46,6 @@ const exceptions = [
   'is-terminal',
 ];
 
-// Astro i18n options
-const i18nOptions = {
-  defaultLocale: 'en',
-  locales: ['en', 'pt'],
-  explicitLocales: {
-    en: 'en-US',
-    pt: 'pt-BR',
-  },
-};
-
 // https://astro.build/config
 export default defineConfig({
   // Your final, deployed URL.
@@ -70,8 +61,8 @@ export default defineConfig({
   redirects: getRedirects(),
   // Configures i18n routing
   i18n: {
-    defaultLocale: i18nOptions.defaultLocale,
-    locales: i18nOptions.locales,
+    defaultLocale: defaultLocale,
+    locales: locales,
   },
   // Set the route matching behavior
   trailingSlash: 'always',
@@ -102,8 +93,8 @@ export default defineConfig({
     }),
     sitemap({
       i18n: {
-        defaultLocale: i18nOptions.defaultLocale,
-        locales: i18nOptions.explicitLocales,
+        defaultLocale: defaultLocale,
+        locales: explicitLocales,
       },
     }),
     robotsTxt({
